feat(security/pipeline): Structure pipeline code
